tzlog: Deregister shmbridge on Hibernate entry
Deregister shmbridge on hibernate entry so that shmbridge
list is cleared for tzlog driver. This is needed because
on hibernate exit tz log driver will again register itself
with shmbridge and this call will fail if list will have old handle.

Tests:
Validated QseeLog and TZlog on Hibernate exit and Quickboot.
(On both secure and Non-secure device).

Change-Id: I37f21efffbc3f6b3b4514bca45f1fbda20b0fcf5
Signed-off-by: Shivangi Kesharwani <quic_skesharw@quicinc.com>